legal-analyzer-microservice/requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
spacy
pydantic
transformers
torch